# src/go/web/public/docs/openapi.yml
openapi: 3.0.0
info:
  description: phenix API
  version: 0.0.0
  title: phenix
  termsOfService: http://swagger.io/terms/
  contact:
    url: https://github.com/sandialabs/sceptre-phenix
  license:
    name: GNU General Public License v3.0
...
paths:
...
  "/roles":
    get:
      tags:
        - Roles
      summary: Get all roles
      operationId: getRoles
      responses:
        "200":
          description: roles
          content:
            application/json:
              schema:
                type: object
    post:
      tags:
        - Roles
      summary: Create a role
      operationId: createRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                policies:
                  type: array
                  items:
                    type: object
      responses:
        "201":
          description: role created
        "400":
          description: invalid request
        "403":
          description: forbidden
  "/roles/{name}":
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - Roles
      summary: Get a role
      operationId: getRole
      responses:
        "200":
          description: role
          content:
            application/json:
              schema:
                type: object
        "403":
          description: forbidden
        "404":
          description: role not found
    patch:
      tags:
        - Roles
      summary: Update a role
      operationId: updateRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                policies:
                  type: array
                  items:
                    type: object
      responses:
        "204":
          description: role updated
        "400":
          description: invalid request
        "403":
          description: forbidden
        "404":
          description: role not found
    delete:
      tags:
        - Roles
      summary: Delete a role
      operationId: deleteRole
      responses:
        "204":
          description: role deleted
        "403":
          description: forbidden
        "404":
          description: role not found
...
servers:
  - url: http://localhost:3000/api/v1
...